updated imports of benchmark helpers, renamed benchmarks, updated imports of benchmark inits
