+++ brainscore_vision/benchmarks/imagenet_c/__init__.py
@@ -1,5 +1,5 @@
 from brainscore_vision import benchmark_registry
-from .imagenet_c import Imagenet_C_Noise, Imagenet_C_Blur, Imagenet_C_Weather, Imagenet_C_Digital
+from .benchmark import Imagenet_C_Noise, Imagenet_C_Blur, Imagenet_C_Weather, Imagenet_C_Digital
 
 benchmark_registry['dietterich.Hendrycks2019-noise-top1'] = Imagenet_C_Noise
 benchmark_registry['dietterich.Hendrycks2019-blur-top1'] = Imagenet_C_Blur
