+++ brainscore_vision/benchmarks/marques2020/benchmarks/schiller1976_benchmark.py
