+++ brainscore_vision/benchmarks/rajalingham2018/benchmarks/public_benchmark.py
@@ -5,5 +5,5 @@ from brainscore_vision.metrics.image_level_behavior import I2n
 class RajalinghamMatchtosamplePublicBenchmark(_DicarloRajalingham2018):
     def __init__(self):
         super(RajalinghamMatchtosamplePublicBenchmark, self).__init__(metric=I2n(), metric_identifier='i2n')
-        self._assembly = LazyLoad(lambda: load_rajalingham2018(access='public'))
+        self._assembly = load_rajalingham2018(access='public')
         self._ceiling_func = lambda: self._metric.ceiling(self._assembly, skipna=True)
